// client/src/app/views/dashboard/model/customSeance.ts
export interface CustomSeance {
  nomFiliere: string;
  nomMatiere: string;
  formatedDate: string;
  formatedDateEnd: string;
  nomProf: string;
  prenomProf: string;
  prenomProf2: string;
  nomSalle: string;
  listEtudiants: string[];
  idSalle: string;
  idGroupe: string;
  idProf: string;
  idEnseignement: string;
}